refactor(console): delete old ajs + replace w/category migration
