+++ gravitee-apim-console-webui/src/index.scss
@@ -45,7 +45,6 @@
   @import 'management/analytics';
   @import 'management/settings/portal-theme/portal-theme';
   @import 'management/settings/groups/group/group';
-  @import 'management/settings/categories/category/category';
   @import 'management/datatable';
   @import 'management/forms';
   @import 'management/stepper';
