+++ gravitee-apim-console-webui/src/management/settings/categories/add-api-to-category-dialog/add-api-to-category-dialog.component.html
