+++ gravitee-apim-console-webui/src/management/settings/categories/categories.component.html
