+++ gravitee-apim-console-webui/src/management/settings/categories/categories.component.scss
@@ -1,4 +1,4 @@
-@use '../../../../../scss/gio-layout' as gio-layout;
+@use '../../../scss/gio-layout' as gio-layout;
 
 :host {
   @include gio-layout.gio-responsive-margin-container;
