--- gravitee-apim-console-webui/src/management/settings/categories/category/_category.scss
@@ -1,8 +1,0 @@
-.new-category-form-input-action {
-  label {
-    margin-left: 25px;
-  }
-  ng-md-icon {
-    top: 0px;
-  }
-}
