+++ gravitee-apim-console-webui/src/management/settings/categories/category/category.component.html
