+++ gravitee-apim-console-webui/src/management/settings/categories/category/category.component.scss
@@ -1,4 +1,4 @@
-@use '../../../../../../scss/gio-layout' as gio-layout;
+@use '../../../../scss/gio-layout' as gio-layout;
 @use 'sass:map';
 @use '@angular/material' as mat;
 @use '@gravitee/ui-particles-angular' as gio;
