--- gravitee-apim-console-webui/src/management/settings/categories/category/delete-api-category.dialog.html
@@ -1,26 +1,0 @@
-<!--
-
-    Copyright (C) 2015 The Gravitee team (http://gravitee.io)
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-            http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<md-dialog aria-label="Detach api '{{api}}'">
-  <md-dialog-content layout-padding>
-    <p>Are you sure you want to detach the API {{api}} ?</p>
-  </md-dialog-content>
-  <md-dialog-actions layout="row">
-    <md-button ng-click="cancel()" md-auto-focus> Cancel </md-button>
-    <md-button ng-click="ok()" class="md-raised md-warn"> Detach API </md-button>
-  </md-dialog-actions>
-</md-dialog>
