+++ gravitee-apim-console-webui/src/management/settings/settings-routing.module.ts
@@ -37,8 +37,6 @@ import { ApiQualityRulesComponent } from './api-quality-rules/api-quality-rules.
 import { EnvironmentNotificationComponent } from './notification/environment-notification.component';
 import { IdentityProvidersComponent } from './identity-providers/identity-providers.component';
 import { PortalSettingsComponent } from './portal-settings/portal-settings.component';
-import { CategoriesNgxComponent } from './categories/migrated/categories-ngx/categories-ngx.component';
-import { CategoryNgxComponent } from './categories/migrated/categories-ngx/category-ngx/category-ngx.component';
 
 import { DocumentationEditPageComponent } from '../../components/documentation/edit-page.component';
 import { DocumentationImportPagesComponent } from '../../components/documentation/import-pages.component';
@@ -136,25 +134,6 @@ export const settingsRoutes: Routes = [
           permissions: {
             anyOf: ['environment-category-r'],
             unauthorizedFallbackTo: '../client-registration-providers',
-          },
-        },
-      },
-      {
-        path: 'categories-ngx',
-        component: CategoriesNgxComponent,
-        data: {
-          permissions: {
-            anyOf: ['environment-category-r'],
-            unauthorizedFallbackTo: '../client-registration-providers',
-          },
-        },
-      },
-      {
-        path: 'categories-ngx/:categoryId',
-        component: CategoryNgxComponent,
-        data: {
-          permissions: {
-            anyOf: ['environment-category-u', 'environment-category-d', 'environment-category-r'],
           },
         },
       },
